Added sample inspection
Fixes #3

The sample inspection feature allows the raw samples to be inspected
when they are fresh, but before they have been added to the summing
variable before finally being averaged. This feature allows samples to
be thrown out, filtered, inspected for certain properties, or any of a
number of different things.

Separated the advanced function features from the normal features when
setting up a channel. This allowed the most often used features to be
quicker to implement in a project, and allow the complex features be
only used on those channels that require it.